Limitation implementation

+++ routes/user.js
@@ -6,25 +6,30 @@ const { Project } = require('../models/Project');
 const { Column } = require('../models/Column');
 const { Task } = require('../models/Task');
 const mongoose = require('mongoose');
+const limitations = require('../limitations');
 
 // Create Projects
 router.post('/projects', isLoggedIn, async (req, res, next) => {
     try {
         let project = new Project(req.body); // Create project instance
         await project.joiValidate(req.body);  // Wait for validation
-        let user = User.findById(req.user._id); // Check user exists (not necessary because middleware already checked but its more safer.)
+        let user = User.findById(req.user._id, { projects: 1 }); // Check user exists (not necessary because middleware already checked but its more safer.)
         if (user === null) { // If user not exists
             req.logout(); // Terminates session for security.
         } else { // If user exists
-            project = await project.save(req.body); // Save project to database
-            await User.findByIdAndUpdate(req.user._id, { "$push": { "projects": project._id } }, { "new": true, "upsert": true }); // Save project to user on database
-            // TODO: After project is created, automatically create "To-Do", "In-Progress", "Done" columns.
-            let todo = await new Column({ "title": "To-Do", "icon": "todo", "project_id": project._id }).save();
-            let inprogress = await new Column({ "title": "In Progress", "icon": "inprogress", "project_id": project._id }).save();
-            let done = await new Column({ "title": "Done", "icon": "todo", "project_id": project._id }).save();
+            if (req.user.projects.length >= limitations.PROJECT_COUNT_PER_USER) {
+                res.status(429).json({ status: false, message: `Due to limitations users cannot create more than ${limitations.PROJECT_COUNT_PER_USER} projects.` })
+            } else {
+                project = await project.save(req.body); // Save project to database
+                await User.findByIdAndUpdate(req.user._id, { "$push": { "projects": project._id } }, { "new": true, "upsert": true }); // Save project to user on database
+                // TODO: After project is created, automatically create "To-Do", "In-Progress", "Done" columns.
+                let todo = await new Column({ "title": "To-Do", "icon": "todo", "project_id": project._id }).save();
+                let inprogress = await new Column({ "title": "In Progress", "icon": "inprogress", "project_id": project._id }).save();
+                let done = await new Column({ "title": "Done", "icon": "todo", "project_id": project._id }).save();
 
-            await Project.findByIdAndUpdate(project._id, { "$push": { "columns": [todo._id, inprogress._id, done._id] } }, { "upsert": true }); // Add column to project on database
-            res.status(200).send({ status: true, message: 'Project successfully created.' });
+                await Project.findByIdAndUpdate(project._id, { "$push": { "columns": [todo._id, inprogress._id, done._id] } }, { "upsert": true }); // Add column to project on database
+                res.status(200).json({ status: true, message: 'Project successfully created.' });
+            }
         }
 
     } catch (err) {
@@ -48,7 +53,7 @@ router.get('/projects', isLoggedIn, async (req, res, next) => {
             }
         } else { // If user didn't specified any project id
             let projects = await Project.find({}, { __v: 0 }); // Get all projects
-            res.status(200).send({ status: true, results: projects });
+            res.status(200).json({ status: true, results: projects });
         }
     } catch (err) {
         res.status(500).json({ status: false, message: err }); // If any error occurs
@@ -62,6 +67,7 @@ router.delete('/projects', isLoggedIn, async (req, res, next) => {
             if (mongoose.Types.ObjectId.isValid(req.query.project_id)) { // Check project id is valid
                 let deletedProject = await Project.deleteOne({ _id: req.query.project_id });
                 if (deletedProject.deletedCount && deletedProject.deletedCount > 0) {
+                    await User.findByIdAndUpdate(req.user._id, { "$pull": { "projects": req.query.project_id } }, { "new": true, "upsert": true });
                     res.status(200).json({ status: true, result: deletedProject });
                 } else {
                     res.status(500).json({ status: true, result: "Error occurred during delete operation" });
@@ -70,7 +76,7 @@ router.delete('/projects', isLoggedIn, async (req, res, next) => {
                 res.status(400).json({ status: false, message: "project_id is malformed." });
             }
         } else { // If user didn't specified any project id
-            res.status(400).send({ status: false, message: "project_id is required." });
+            res.status(400).json({ status: false, message: "project_id is required." });
         }
     } catch (err) {
         res.status(500).json({ status: false, message: err }); // If any error occurs
@@ -93,7 +99,7 @@ router.put('/projects', isLoggedIn, async (req, res, next) => {
                         res.status(404).json({ status: false, message: "Project is not exists." });
                     } else { // If project exists
                         project = await Project.findByIdAndUpdate(req.body.project_id, { $set: req.body }, { new: true });
-                        res.status(200).send({ status: true, message: 'Project successfully updated.' });
+                        res.status(200).json({ status: true, message: 'Project successfully updated.' });
                     }
                 }
 
@@ -101,7 +107,7 @@ router.put('/projects', isLoggedIn, async (req, res, next) => {
                 res.status(400).json({ status: false, message: "project_id is malformed." });
             }
         } else { // If user didn't specified any project id
-            res.status(400).send({ status: false, message: "project_id is required." });
+            res.status(400).json({ status: false, message: "project_id is required." });
         }
     } catch (err) {
         res.status(500).json({ status: false, message: err }); // If any error occurs
@@ -118,9 +124,13 @@ router.post('/columns', isLoggedIn, async (req, res, next) => {
             if (project === null) { // If project is not exists.
                 res.status(404).json({ status: false, message: "Project is not exists." });
             } else { // If project exists
-                column = await column.save(req.body); // Save column to database
-                await Project.findByIdAndUpdate(req.body.project_id, { "$push": { "columns": column._id } }, { "upsert": true }); // Add column to project on database
-                res.status(200).send({ status: true, message: 'Column successfully created.' });
+                if (project.columns.length >= limitations.COLUMN_COUNT_PER_PROJECT) {
+                    res.status(429).json({ status: false, message: `Due to limitations projects cannot contains more than ${limitations.COLUMN_COUNT_PER_PROJECT} columns.` })
+                } else {
+                    column = await column.save(req.body); // Save column to database
+                    await Project.findByIdAndUpdate(req.body.project_id, { "$push": { "columns": column._id } }, { "upsert": true }); // Add column to project on database
+                    res.status(200).json({ status: true, message: 'Column successfully created.' });
+                }
             }
         } else {
             res.status(400).json({ status: false, message: "project_id is undefined or malformed." }); // Project id is undefined or malformed
@@ -146,7 +156,7 @@ router.put('/columns', isLoggedIn, async (req, res, next) => {
                         res.status(404).json({ status: false, message: "Column is not exists." });
                     } else { // If column exists
                         column = await Column.findByIdAndUpdate(req.body.column_id, { $set: req.body }, { new: true });
-                        res.status(200).send({ status: true, message: 'Column successfully updated.' });
+                        res.status(200).json({ status: true, message: 'Column successfully updated.' });
                     }
                 }
 
@@ -154,7 +164,7 @@ router.put('/columns', isLoggedIn, async (req, res, next) => {
                 res.status(400).json({ status: false, message: "column_id is malformed." });
             }
         } else { // If user didn't specified any column id
-            res.status(400).send({ status: false, message: "column_id is required." });
+            res.status(400).json({ status: false, message: "column_id is required." });
         }
     } catch (err) {
         res.status(500).json({ status: false, message: err }); // If any error occurs
@@ -165,18 +175,27 @@ router.put('/columns', isLoggedIn, async (req, res, next) => {
 router.delete('/columns', isLoggedIn, async (req, res, next) => {
     try {
         if (req.query.column_id) { // Check user specified task id
-            if (mongoose.Types.ObjectId.isValid(req.query.column_id)) { // Check task id is valid
-                let deletedColumn = await Column.deleteOne({ _id: req.query.column_id });
-                if (deletedColumn.deletedCount && deletedColumn.deletedCount > 0) {
-                    res.status(200).json({ status: true, result: deletedColumn });
-                } else {
-                    res.status(500).json({ status: true, result: "Error occurred during delete operation" });
+            if (req.query.project_id) {
+                if (mongoose.Types.ObjectId.isValid(req.query.column_id)) { // Check task id is valid
+                    if (mongoose.Types.ObjectId.isValid(req.query.project_id)) {
+                        let deletedColumn = await Column.deleteOne({ _id: req.query.column_id });
+                        if (deletedColumn.deletedCount && deletedColumn.deletedCount > 0) {
+                            await Project.findByIdAndUpdate(req.query.project_id, { "$pull": { "columns": req.query.column_id } }, { "new": true, "upsert": true });
+                            res.status(200).json({ status: true, result: deletedColumn });
+                        } else {
+                            res.status(500).json({ status: true, result: "Error occurred during delete operation" });
+                        }
+                    } else {
+                        res.status(400).json({ status: false, message: "project_id is malformed." });
+                    }
+                } else { // If task is malformed
+                    res.status(400).json({ status: false, message: "column_id is malformed." });
                 }
-            } else { // If task is malformed
-                res.status(400).json({ status: false, message: "column_id is malformed." });
+            } else {
+                res.status(400).json({ status: false, message: "project_id is required." });
             }
         } else { // If user didn't specified any task id
-            res.status(400).send({ status: false, message: "column_id is required." });
+            res.status(400).json({ status: false, message: "column_id is required." });
         }
     } catch (err) {
         res.status(500).json({ status: false, message: err }); // If any error occurs
@@ -199,7 +218,7 @@ router.get('/columns', isLoggedIn, async (req, res, next) => {
             }
         } else {
             let columns = await Column.find({}, { __v: 0 });
-            res.status(200).send({ status: true, results: columns });
+            res.status(200).json({ status: true, results: columns });
         }
     } catch (err) {
         res.status(500).json({ status: false, message: err }); // If any error occurs
@@ -218,7 +237,7 @@ router.post('/tasks', isLoggedIn, async (req, res, next) => {
             } else {
                 task = await task.save(req.body);
                 await Column.findByIdAndUpdate(req.body.column_id, { "$push": { "tasks": task._id } }, { "new": true, "upsert": true });
-                res.status(200).send({ status: true, message: 'Task successfully created.' });
+                res.status(200).json({ status: true, message: 'Task successfully created.' });
             }
 
         } else {
@@ -233,18 +252,27 @@ router.post('/tasks', isLoggedIn, async (req, res, next) => {
 router.delete('/tasks', isLoggedIn, async (req, res, next) => {
     try {
         if (req.query.task_id) { // Check user specified task id
-            if (mongoose.Types.ObjectId.isValid(req.query.task_id)) { // Check task id is valid
-                let deletedTask = await Task.deleteOne({ _id: req.query.task_id });
-                if (deletedTask.deletedCount && deletedTask.deletedCount > 0) {
-                    res.status(200).json({ status: true, result: deletedTask });
-                } else {
-                    res.status(500).json({ status: true, result: "Error occurred during delete operation" });
+            if (req.query.column_id) {
+                if (mongoose.Types.ObjectId.isValid(req.query.task_id)) { // Check task id is valid
+                    if (mongoose.Types.ObjectId.isValid(req.query.column_id)) {
+                        let deletedTask = await Task.deleteOne({ _id: req.query.task_id });
+                        if (deletedTask.deletedCount && deletedTask.deletedCount > 0) {
+                            await Column.findByIdAndUpdate(req.query.column_id, { "$pull": { "tasks": req.query.task_id } }, { "new": true, "upsert": true });
+                            res.status(200).json({ status: true, result: deletedTask });
+                        } else {
+                            res.status(500).json({ status: true, result: "Error occurred during delete operation" });
+                        }
+                    } else {
+                        res.status(400).json({ status: false, message: "column_id is malformed." });
+                    }
+                } else { // If task is malformed
+                    res.status(400).json({ status: false, message: "task_id is malformed." });
                 }
-            } else { // If task is malformed
-                res.status(400).json({ status: false, message: "task_id is malformed." });
+            } else {
+                res.status(400).json({ status: false, message: "column_id is required." });
             }
         } else { // If user didn't specified any task id
-            res.status(400).send({ status: false, message: "task_id is required." });
+            res.status(400).json({ status: false, message: "task_id is required." });
         }
     } catch (err) {
         res.status(500).json({ status: false, message: err }); // If any error occurs
